重构数据库模型和迁移脚本- 重新定义了用户、房间、单独用户和角色信息的数据库模型 - 添加了新的关联表以支持用户与单独/房间角色的多对多关系 - 修改了数据库迁移脚本以反映新的模型结构 - 更新了实体模块以包含新的关联实体 -调整了用户表的主键类型从i64改为i32

+++ backend/src/sql/user.rs
@@ -23,7 +23,7 @@ impl DataBase for UserDB {
     }
 }
 
-type PrimaryKey = i64;
+type PrimaryKey = i32;
 
 impl UserDB {
     pub async fn from_cfg(config: &DataBaseConfig) -> Result<Self, Error> {
